requirements.txt
asgiref==3.5.2
Django==4.1.1
django-crispy-forms==1.14.0
django-filter==22.1
django-widget-tweaks==1.4.12
Pillow==9.2.0
psycopg2==2.9.5
sqlparse==0.4.2
tzdata==2022.2
